requirements.txt
Jinja2==2.11.2
PyYAML==5.3.1
python-slugify==4.0.1
